code/requirements.txt
ambiance==1.3.0
black==22.1.0
bleach==4.1.0
build==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
data-handler-csv==0.0.3
docutils==0.18.1
idna==3.3
importlib-metadata==4.11.3
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.0
matplotlib==3.4.3
mypy-extensions==0.4.3
numpy==1.21.3
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
pep517==0.12.0
Pillow==8.4.0
pkginfo==1.8.2
platformdirs==2.5.1
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
readme-renderer==34.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
scipy==1.7.1
SecretStorage==3.3.1
six==1.16.0
tomli==2.0.1
tqdm==4.63.0
twine==3.8.0
typing-extensions==4.1.1
urllib3==1.26.9
webencodings==0.5.1
zipp==3.7.0
